add react router dom route to home
